pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "numinous"
version = "0.1.0"
description = "The ground the mesh floats in."
requires-python = ">=3.11"
dependencies = [
    "manifold-mesh",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["numinous*"]
